requirements.txt
langchain
langchain_openai
langchain_community 
python-dotenv 
pymysql 
PyAthena 
streamlit
fastapi